Add patches to patch the ungoogled-chromium patches
This should make fixing the ungoogled-chromium patches a lot easier and
eliminate the need to maintain modified copies of them locally.

Should any of the patches fail to apply, they can be reconstructed using
diff, but only after the ungoogled-chromium patches have been applied to
the Chromium source. Otherwise, the patch will be empty.