Patch ungoogled-chromium patches during build
Normally, the build system in use for FreeBSD Ports applies any existing
patches it can find automatically, but this is limited to one patch
directory. The patches within that directory are then applied to the
Chromium source code, which has its own subdirectory within the work
directory.

The ungoogled-chromium patches are in its own subdirectory, and they
need to be patched before they can be used or else the build fails, so
we will just apply them manually during the pre-configure phase.